requirements.txt
Django==1.9.7
Pillow==3.3.0
django-debug-toolbar==1.4
djangorestframework==3.3.3
rest-condition==1.0.2
sqlparse==0.1.19
